website is done

+++ middlewares/deleteAdventure.js
@@ -5,6 +5,9 @@
  */
 module.exports = (objRepo) => {
     return (req, res, next)=>{
+        if (!res.locals.adventure._id) {
+            return res.redirect("/adventures");
+        }
         return res.locals.adventure.deleteOne().then(() => {
             return res.redirect("/adventures");
         }).catch(next);
